feat: update validation and db data type

+++ app/Http/Requests/Movie/ComingSoonMovie/StoreRequest.php
@@ -16,7 +16,7 @@ class StoreRequest extends BaseRequest
         return [
             'title' => ['required', 'string', 'unique:coming_soon_movies'],
             'plot' => ['required', 'string'],
-            'duration_in_minutes' => ['required', 'integer'],
+            'duration_in_minutes' => ['required', 'numeric'],
             'age_restriction' => ['required', 'integer'],
             'country' => ['required', 'string'],
             'language' => ['required', 'string'],
