+++ app/Http/Requests/Movie/Movie/StoreRequest.php
@@ -18,7 +18,7 @@ class StoreRequest extends BaseRequest
             'plot' => ['required', 'string'],
             'year_of_release' => ['required', 'integer'],
             'date_of_release' => ['required', 'date'],
-            'duration_in_minutes' => ['required', 'integer'],
+            'duration_in_minutes' => ['required', 'numeric'],
             'age_restriction' => ['required', 'integer'],
             'country' => ['required', 'string'],
             'language' => ['required', 'string'],
